Updates for the UI to grab the settings all the once
