+++ imbi/endpoints/ui/__init__.py
@@ -4,8 +4,7 @@ System Reports
 """
 from tornado import web
 
-from . import automations, authentication, groups, index, metadata, \
-    settings, user
+from . import automations, authentication, groups, index, settings, user
 
 IndexRequestHandler = index.IndexRequestHandler
 
@@ -13,8 +12,7 @@ URLS = automations.URLS + [
     web.url(r'^/ui/', index.IndexRequestHandler),
     web.url(r'^/ui/login$', authentication.LoginRequestHandler),
     web.url(r'^/ui/logout$', authentication.LogoutRequestHandler),
-    web.url(r'^/ui/metadata$', metadata.RequestHandler),
-    web.url(r'^/ui/settings$', settings.SettingsRequestHandler),
+    web.url(r'^/ui/settings$', settings.RequestHandler),
     web.url(r'^/ui/user$', user.UserRequestHandler),
     web.url(r'^/ui/.*$', index.IndexRequestHandler)
 ]
